Simplify word wrap and scrolling; vertical scrolling is now tracked in (Lines, Inner Lines). When word wrap is disabled, (inner lines) is always 0. Otherwise it indicates how far along the current line we are.
When scrolling and word wrap is enabled, the WrappedTextIterator is used
to determine whether (line) or (inner line) should be
incremented/decremented.

The scrollbar now always scrolls in full line
numbers, and no longer scrolls partial lines, even when word wrap is
enabled. This is done so we no longer need to parse the entire file to
determine scroll bar height/offset when word wrap is enabled.