To avoid many expensive buffer updates to start_line, both "start_line" and "cumulative_width" are now not guaranteed to be correct while an operation is performed (i.e. in the middle of a PerformOperation call, they should no longer be used during this time).
To cover the same use cases that they used to cover, TextBuffer now has
three extra pieces of information: "index" (index in buffer vector), "width" (cumulative width of all lines in the buffer)  and "line_count" (amount of lines in the buffer).
These are always updated and should always be correct.

The start_line and cumulative_width properties are restored to a correct
state in the InvalidateBuffers() call, which is always called after
calling PerformOperation().