Add replay functionality.
Replays can be created by calling PGGlobalReplayManager::Initialize.
Replays save all the events that happen in a ControlManager, as well as
all the contents of files that are opened, while Panther is running.
The main purpose of these replays is to create test cases and for crash
reporting: replays can be created with a specific input and an expected
output, and ran after changes are made to the code.